Update method used to locate items
When a directory change notification is received, the internal data
structures associated with the changed item need to be found. This was
previously done by finding the item in the listview, then using that
information to locate the internal data structures. However, if the
item wasn't in the listview (because it had been filtered out, for
example), the item wouldn't be found and would never get updated.

To resolve this, the internal index associated with an item is now
determined directly.